quotes/quotes/requirements.txt
Flask==2.2.2
psycopg2-binary==2.9.6
Scrapy==2.8.0
